src/main/java/dev/drf/leetcode/pacific/atlantic/water/flow/PacificAtlanticWaterFlowProblem.java
package dev.drf.leetcode.pacific.atlantic.water.flow;

import dev.drf.leetcode.Problem;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

/**
 * https://leetcode.com/problems/pacific-atlantic-water-flow/
 * Used solution:
 * https://www.youtube.com/watch?v=s-VkcjHqkGI
 */
public class PacificAtlanticWaterFlowProblem implements Problem {
    public List<List<Integer>> pacificAtlantic(int[][] heights) {
        List<List<Integer>> result = new ArrayList<>();

        int rows = heights.length, cols = heights[0].length;
        boolean[][] pacific = new boolean[rows][cols];
        boolean[][] atlantic = new boolean[rows][cols];

        for (int i = 0; i < cols; i++) {
            dfs(heights, 0, i, Integer.MIN_VALUE, pacific);
            dfs(heights, rows - 1, i, Integer.MIN_VALUE, atlantic);
        }

        for (int i = 0; i < rows; i++) {
            dfs(heights, i, 0, Integer.MIN_VALUE, pacific);
            dfs(heights, i, cols - 1, Integer.MIN_VALUE, atlantic);
        }

        for (int i = 0; i < rows; i++) {
            for (int j = 0; j < cols; j++) {
                if (pacific[i][j] && atlantic[i][j]) {
                    result.add(Arrays.asList(i, j));
                }
            }
        }
        return result;
    }

    private void dfs(int[][] heights,
                     int i,
                     int j,
                     int prev,
                     boolean[][] ocean) {
        if (i < 0 || i >= ocean.length
                || j < 0 || j >= ocean[0].length) {
            return;
        }
        if (heights[i][j] < prev
                || ocean[i][j]) {
            return;
        }

        ocean[i][j] = true;
        dfs(heights, i, j + 1, heights[i][j], ocean);
        dfs(heights, i, j - 1, heights[i][j], ocean);
        dfs(heights, i + 1, j, heights[i][j], ocean);
        dfs(heights, i - 1, j, heights[i][j], ocean);
    }
}
